Refactoring: moved ISearchLine model into ILine
