+++ src/components/shared/searchView/SearchResults.tsx
@@ -2,8 +2,8 @@ import { reaction, IReactionDisposer } from 'mobx';
 import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { matchPath } from 'react-router';
+import { ISearchLine } from '~/models/ILine';
 import { INodeBase } from '~/models/INode';
-import ISearchLine from '~/models/searchModels/ISearchLine';
 import Navigator from '~/routing/navigator';
 import subSites from '~/routing/subSites';
 import LineService from '~/services/lineService';
