--- src/models/searchModels/ISearchLine.ts
@@ -1,8 +1,0 @@
-import TransitType from '~/enums/transitType';
-import ISearchLineRoute from './ISearchLineRoute';
-
-export default interface ISearchLine {
-    id: string;
-    transitType: TransitType;
-    routes: ISearchLineRoute[];
-}
