--- src/models/searchModels/ISearchLineRoute.ts
@@ -1,6 +1,0 @@
-export default interface ISearchLineRoute {
-    id: string;
-    name: string;
-    isUsedByRoutePath: boolean;
-    date?: Date;
-}
