+++ src/stores/lineStore.ts
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 import { action, computed, observable, reaction } from 'mobx';
 import { ILine, IRoute } from '~/models';
-import ISearchLine from '~/models/searchModels/ISearchLine';
+import { ISearchLine } from '~/models/ILine';
 import lineValidationModel, {
-    ILineValidationModel
+    ILineValidationModel,
 } from '~/models/validationModels/lineValidationModel';
 import { IValidationResult } from '~/validation/FormValidator';
 import NavigationStore from './navigationStore';
@@ -87,14 +87,14 @@ class LineStore {
                         if (this.isLineAlreadyFound(lineId)) {
                             const validationResult: IValidationResult = {
                                 isValid: false,
-                                errorMessage: `Linja ${lineId} on jo olemassa.`
+                                errorMessage: `Linja ${lineId} on jo olemassa.`,
                             };
                             return validationResult;
                         }
                         return;
-                    }
-                ]
-            }
+                    },
+                ],
+            },
         };
 
         this._validationStore.init(line, lineValidationModel, customValidatorMap);
